requirements.txt
google-generativeai
python-dotenv
pyperclip
keyboard
PyYAML